Refactored some code and Made specific page for each sidebar item
