+++ frontend/src/components/SideBar.jsx
@@ -17,32 +17,29 @@ const Sidebar = ({ setSelectedCategory }) => {
     <div className="w-24 bg-gray-800 text-white h-screen flex flex-col items-center p-4 justify-between border-r-2 border-gray-700">
       <div>
         <div className="md:flex md:items-center md:gap-12 text-[#5e17eb] font-semibold">
-          <a
-            className="flex items-end gap-1"
-            onClick={() => setSelectedCategory("LandingPage")}
-          >
+          <Link className="flex items-end gap-1" to="/home">
             <img src={Logo} alt="Logo" className="h-14 w-auto" />
-          </a>
+          </Link>
         </div>
         <div className="flex flex-col items-center space-y-4 mt-6">
-          <div
+          <Link
+            to="/servers"
             className="w-11 h-11 flex items-center justify-center rounded-full bg-gray-700 hover:bg-gray-600 cursor-pointer"
-            onClick={() => setSelectedCategory("server")}
           >
             <FaServer className="text-xl text-white" />
-          </div>
-          <div
+          </Link>
+          <Link
+            to="/groups"
             className="w-11 h-11 flex items-center justify-center rounded-full bg-gray-700 hover:bg-gray-600 cursor-pointer"
-            onClick={() => setSelectedCategory("group")}
           >
             <FaUserFriends className="text-xl text-white" />
-          </div>
-          <div
+          </Link>
+          <Link
             className="w-11 h-11 flex items-center justify-center rounded-full bg-gray-700 hover:bg-gray-600 cursor-pointer"
-            onClick={() => setSelectedCategory("dm")}
+            to="/directmessages"
           >
             <FaPaperPlane className="text-xl text-white" />
-          </div>
+          </Link>
           <div
             className="w-11 h-11 flex items-center justify-center rounded-full bg-gray-700 hover:bg-gray-600 cursor-pointer"
             onClick={() => setIsSettingsOpen(true)}
